perf(eth): move ethereum handler into etheruem listener module
