+++ src/listener/eth/index.ts
@@ -2,17 +2,19 @@ import { BlockchainState } from "./BlockchainState";
 import { EventParser } from "./EventParser";
 import { blockInDB, logInDB } from "./DB";
 import { localConfig } from "./Config";
-import { LogType } from "./types";
+import { ITx } from "../../types";
+import { log } from "../../util";
 
 const blockchainState = new BlockchainState();
 const eventParser = new EventParser();
 
-export function listen(config: any, callback: (
-    tx: string, type: LogType, blockNumber: number,
-) => void) {
+export function listen(config: any, queue: ITx[]) {
     localConfig.setConfig(config);
     blockInDB.setParsedEventBlockNumber(localConfig.info.START_BLOCK_NUMBER);
-    logInDB.setCallback(callback);
+    logInDB.setCallback(async (tx: string, ty: string, blockNumber: number) => {
+        log.trace(`Find darwinia ${ty} tx ${tx} in block ${blockNumber}`);
+        queue.push({ blockNumber, tx, ty: ty === "bank" ? "Deposit" : "Token" });
+    });
     blockchainState.getState().then(() => {
         eventParser.start(blockchainState);
     })
